This is a breaking change to the API and is designed to extend support to arbitrary *monkey extensions by granting both GM_ and GM. namespaced functions, and interpolating the correct one to use depending on the scriptHandler.
On init, the script tests for the presence of a valid GM_info object and
sets the gmPrefix variable to the string "GM_". If it returns null, it
sets it to "GM."

kes.user.js and mods call the function safeGM(), which sanitizes the
request and checks gmPrefix for the desired namespace. The argument is
the bare form of the GM function without the prefix, e.g., "setValue",
and is a string.

The arguments are passed verbatim through safeGM, which chooses from one
of two objects to find the desired function name mapping. From there,
the function is called as if it were being invoked normally.

Both the GM_ and GM. form of the grants have to be included in the
header, and the internal GM functions being exposed through safeGM must
be explicitly defined in the object contained within safeGM().

Currently, the following have been added. The majority are helper
functions for kes.user.js itself, with a few being used by mods. It is
assumed that mods generally do not need to make heavy use of GM internal
functions.

addStyle
getResourceText
xmlhttpRequest
info
getValue
setValue
getResourceText
setClipboard

Observations on variance between GM_ and GM. functions:
- In the GM_ variant, getValue and getResourceText return a string; in
  the GM. variant, they return a promise. Therefore, an additional
conditional logic is used to parse the promise response in the event
that GM. is being used.
- The GM_ spec defines the XML request function as xmlhttpRequest, but
  the GM. spec defines it as xmlHttpRequest. This must be notated
exactly when including and declaring the functions. For the sake of
sanity, safeGM maps both respective types to the function name
xmlhttpRequest.