fix(fix-codeblocks): lines with multiple erroneous tags are now fully sanitized
Previously the mod assumed that each line gets the span closing and opening tags prepended and that's it. But it's actually possible for there to be multiple erroneous tags per line, as it seems to be more some kind of code formatting. This commit fixes this oversight and replaces erroneous tags on the whole line, not just the start.