fix(improved_collapsible_comments): made the description more truthful
The KES version only supports collapsing comments by clicking their line and the icon next to it. The description however kept the claim from the original userscript that clicking the header would also cause the comment to collapse. This isn't true for the KES version and so I removed that claim.