refactor: moved omni's magazine loading out of the mod
In this commit I moved just the user magazine fetching logic out of the omni(bar) mod and into the safeGM file (for now, as that's where the others are too). This is then directly used by both omni and thread_checkmarks, so a dependency between them isn't necessary anymore.