feat: add anonymize checkbox
Adds an "anonymize" checkbox to the debug bar. This replaces usernames,
magazine names, thread titles, short descriptions, and the logged in
username with dummy text for the purpose of mocking the appearance of
mods without leaking potentially sensitive information. This is going to
be used to expedite completion of #204 when taking screenshots. When
unticking the checkbox, the page reloads.